requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
compressed-tensors==0.8.0
contourpy==1.3.0
cycler==0.12.1
datasets==3.1.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.115.5
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.9.0
gguf==0.10.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
interegular==0.3.3
Jinja2==3.1.4
jiter==0.8.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lark==1.2.2
llvmlite==0.43.0
lm-format-enforcer==0.10.9
MarkupSafe==3.0.2
matplotlib==3.9.3
mistral_common==1.5.1
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.55.3
opencv-python-headless==4.10.0.84
outlines==0.0.46
packaging==24.2
pandas==2.2.3
partial-json-parser==0.2.1.1.post4
pillow==10.4.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.0
propcache==0.2.1
protobuf==5.29.0
psutil==6.1.0
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==18.1.0
pycountry==24.6.1
pydantic==2.10.2
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.39.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.21.0
safetensors==0.4.5
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tiktoken==0.7.0
tokenizers==0.20.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.46.3
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
vllm==0.6.4.post1
watchfiles==1.0.0
websockets==14.1
xformers==0.0.28.post3
xxhash==3.5.0
yarl==1.18.0
zipp==3.21.0
